Move elkm1 base entity to separate module
