+++ homeassistant/components/elkm1/binary_sensor.py
@@ -12,7 +12,8 @@ from homeassistant.components.binary_sensor import BinarySensorEntity
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import ElkAttachedEntity, ElkEntity, ElkM1ConfigEntry
+from . import ElkM1ConfigEntry
+from .entity import ElkAttachedEntity, ElkEntity
 
 
 async def async_setup_entry(
